requirements.txt
python-dotenv
datasets
matplotlib
anthropic
openai
transformers